switched to using EnhancedWordAligner and removed the useAlignmentSuggestions hook from WordAlignerDialog since that is managed in EnhancedWordAligner
