+++ package.json
@@ -45,7 +45,7 @@
     "bible-reference-rcl": "1.4.1",
     "core-js": "^3.8.3",
     "deep-equal": "^2.0.5",
-    "enhanced-word-aligner-rcl": "1.1.0-beta.3",
+    "enhanced-word-aligner-rcl": "1.1.0-beta.4",
     "gitea-react-toolkit": "2.4.0",
     "localforage": "^1.9.0",
     "lodash.clonedeep": "^4.5.0",
